+++ react/Example.tsx
@@ -1,10 +1,17 @@
 import React from 'react'
 import AddressForm from './AddressForm'
-import { sampleAddress as address } from './mocks/CountryDescriptions'
+import LocationSelect from './LocationSelect'
+import LocationCountry from './LocationCountry'
+import {
+  sampleAddress1 as address1,
+  sampleAddress2 as address2,
+} from './mocks/CountryDescriptions'
 import { AddressContextProvider } from 'vtex.address-context/AddressContext'
 import { LocaleSwitcher } from 'vtex.locale-switcher'
 
-const Example: StorefrontFunctionComponent<{}> = () => {
+const countryList = ['BRA', 'KOR', 'ARG', 'BOL']
+
+const Example1: StorefrontFunctionComponent<{}> = () => {
   return (
     <div>
       <h2>AddressForm:</h2>
@@ -13,11 +20,39 @@ const Example: StorefrontFunctionComponent<{}> = () => {
   )
 }
 
+const Example2: StorefrontFunctionComponent<{}> = () => {
+  return (
+    <div>
+      <h2>LocationSelect</h2>
+      <LocationSelect />
+    </div>
+  )
+}
+
+const Example3: StorefrontFunctionComponent<{}> = () => {
+  return (
+    <div>
+      <h2>LocationCountry</h2>
+      <LocationCountry />
+    </div>
+  )
+}
+
 const ExampleWrapper = () => (
-  <AddressContextProvider address={address} countries={['BRA', 'KOR']}>
-    <LocaleSwitcher />
-    <Example />
-  </AddressContextProvider>
+  <div>
+    <AddressContextProvider address={address1} countries={countryList}>
+      <LocaleSwitcher />
+      <Example1 />
+    </AddressContextProvider>
+    <AddressContextProvider address={address2} countries={countryList}>
+      <LocaleSwitcher />
+      <Example2 />
+    </AddressContextProvider>
+    <AddressContextProvider address={address2} countries={countryList}>
+      <LocaleSwitcher />
+      <Example3 />
+    </AddressContextProvider>
+  </div>
 )
 
 export default ExampleWrapper