+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- LocationCountry (that uses country from store settings as first option)
+- LocationSelect
+- Bolivia country JSON
+- Initial styling to `AddressForm` (max/min width depending on fantasy name)
+
+### Removed
+
+- `size` property from country JSON objects
+
 ## [0.3.0] - 2020-02-17
 
 ### Added